pyproject.toml
[project]
name = "llmevalkit"
version = "0.1.0"
description = "Universal LLM eval harness — any model, any benchmark, one runner."
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.27",
    "rich>=13.7",
    "openai>=1.50",
    "pydantic>=2.8",
    "typer>=0.12",
]

[project.optional-dependencies]
shim = ["fastapi>=0.115", "uvicorn>=0.32", "transformers>=4.45", "torch"]
vllm = ["vllm>=0.6"]
all = ["fastapi", "uvicorn", "transformers", "torch", "vllm"]

[project.scripts]
llmevalkit = "llmevalkit.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["llmevalkit"]
